Refactor "Test" tab and improve variable test value editing UI
Renamed `TestRunTab` to `TestTab` and updated related references for consistency. Enhanced `EditVariableTestValueDialog` by replacing redundant fields with a new `DataPanel` for streamlined variable data display. Improved initialization logic to populate `DataPanelModel` with variable metadata. Removed unused imports for cleaner code structure.
