+++ src/modules/Elsa.Studio.Workflows/Components/WorkflowDefinitionEditor/Components/WorkflowProperties/Tabs/Variables/Components/EditVariableTestValueDialog.razor.cs
@@ -1,5 +1,6 @@
 using Elsa.Api.Client.Resources.WorkflowDefinitions.Extensions;
 using Elsa.Api.Client.Resources.WorkflowDefinitions.Models;
+using Elsa.Studio.Models;
 using Elsa.Studio.Workflows.Components.WorkflowDefinitionEditor.Components.WorkflowProperties.Tabs.Variables.Models;
 using Microsoft.AspNetCore.Components;
 using Microsoft.AspNetCore.Components.Forms;
@@ -13,6 +14,7 @@ namespace Elsa.Studio.Workflows.Components.WorkflowDefinitionEditor.Components.W
 public partial class EditVariableTestValueDialog
 {
     private readonly VariableTestValueModel _model = new();
+    private DataPanelModel _variableData = new();
     private EditContext _editContext = null!;
 
     /// <summary>
@@ -24,18 +26,23 @@ public partial class EditVariableTestValueDialog
     /// The variable to edit. If null, a new variable will be created.
     /// </summary>
     [Parameter] public Variable Variable { get; set; } = null!;
+
     [CascadingParameter] private IMudDialogInstance MudDialog { get; set; } = null!;
 
     /// <inheritdoc />
     protected override void OnParametersSet()
     {
-        // Instantiate the edit context first, so that it is available when rendering (which happens as soon as we call an async method on the next lines). 
         _editContext = new(_model);
-        
         _model.VariableId = Variable.Id;
         _model.TestValue = WorkflowDefinition.GetVariableTestValue(Variable.Id)?.ToString() ?? "";
+
+        _variableData = new();
+        _variableData.Add("Variable ID", Variable.Id);
+        _variableData.Add("Name", Variable.Name);
+        _variableData.Add("Type", Variable.TypeName);
+        _variableData.Add("Default Value", Variable.Value?.ToString());
     }
-    
+
     private Task OnCancelClicked()
     {
         MudDialog.Cancel();
